+++ CHANGELOG.md
@@ -37,3 +37,4 @@
   - Indexer now uses Map in place of Object, ( performance improivement )
   - nodes internal structure changed from object, to Map ( performacne improvement )
   - Created Common method for node highlighting
+  - Converted link/index/cache to Map from Object ( performance improvement )